Move FakeBoard file to boards directory
